LPS-63475 moves *Keys and *Constants files to contacts-api
